Bitness fixes
Summary:
Fix instances of unsafe shortening in proxygen.

Test Plan: compiled

Reviewed By: cgheorghe@fb.com

Subscribers: trunkagent, bmatheny, benyluo, ranjeeth, subodh, shikong, fma, kmdent

FB internal diff: D1725037

Tasks: 5538843

Signature: t1:1725037:1418073658:0f2d5daa2d8a2550e0bd9d367dcacae030102bf7

+++ proxygen/lib/http/session/HTTPTransaction.h
@@ -320,9 +320,9 @@ class HTTPTransaction :
 
     virtual void headerBytesReceived(const HTTPHeaderSize& size) noexcept = 0;
 
-    virtual void bodyBytesGenerated(uint32_t nbytes) noexcept = 0;
+    virtual void bodyBytesGenerated(size_t nbytes) noexcept = 0;
 
-    virtual void bodyBytesReceived(uint32_t size) noexcept = 0;
+    virtual void bodyBytesReceived(size_t size) noexcept = 0;
 
     virtual ~TransportCallback() {};
   };
